Lockscreen gestures -
Settings pages: allows user to define custom gestures with
specific activities or shortcuts launched from the lock screen.

Also, allows the user to adjust the sensitivity of the gesture
recognition (a lower number is more forgiving but makes mis-reads
occur more often).

User can also choose to show or hide the gesture 'trail' when
swiping a gesture.

Depends on change 539 and 540